+++ src/components/index.js
@@ -1,9 +1,6 @@
-// Keep in alphabetical order
-// IDEA: Load components dynamically via Weback's req.context?
-
 export { default as App } from './App/App';
-export { default as Avatar } from './Avatar/Avatar';
 export { default as Auth } from './Auth/Auth';
+export { default as Avatar } from './Avatar/Avatar';
 export { default as Board } from './Board/Board';
 export { default as Button } from './Button/Button';
 export { default as InputArea } from './InputArea/InputArea';