+++ src/containers/Message/Message.js
@@ -3,6 +3,7 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { Message as MessageModel } from '../../models';
 import { actions as messageActions } from '../../redux/message';
+import { actions as messagesActions } from '../../redux/messages';
 import { Message as MessageComponent } from '../../components';
 import s from '../../rootSelectors';
 
@@ -11,11 +12,12 @@ class Message extends Component {
     const { message, currentUserId } = this.props;
     return (
       <MessageComponent
+        loading={message.loading}
         text={message.text}
         noIcons={currentUserId === null}
         belongsToCurrentUser={message.author === currentUserId}
         isChild={message.parentId !== null}
-        handleDeleteClick={() => this.props.clickedDelete(message.id)}
+        handleDeleteClick={() => this.props.delete(message.id)}
         handleEditClick={() => this.props.clickedEdit(message)}
         handleRespondClick={() => this.props.clickedRespond(message)}
       />
@@ -26,9 +28,9 @@ class Message extends Component {
 Message.propTypes = {
   message: PropTypes.instanceOf(MessageModel).isRequired,
   currentUserId: PropTypes.number,
-  clickedDelete: PropTypes.func.isRequired,
   clickedEdit: PropTypes.func.isRequired,
   clickedRespond: PropTypes.func.isRequired,
+  delete: PropTypes.func.isRequired,
 };
 
 const mapStateToProps = state => ({
@@ -38,8 +40,8 @@ const mapStateToProps = state => ({
 export default connect(
   mapStateToProps,
   {
-    clickedDelete: messageActions.clickedDelete,
     clickedEdit: messageActions.clickedEdit,
     clickedRespond: messageActions.clickedRespond,
+    delete: messagesActions.delete,
   }
 )(Message);