+++ src/containers/index.js
@@ -1,5 +1,3 @@
-// Keep in alphabetical order
-
 export { default as Auth } from './Auth/Auth';
 export { default as Board } from './Board/Board';
 export { default as InputArea } from './InputArea/InputArea';