+++ src/models/Message.js
@@ -4,7 +4,8 @@ const MessageRecord = Immutable.Record({
   id: null,
   text: '',
   author: null,
-  parentId: null
+  parentId: null,
+  loading: false,
 });
 
 class Message extends MessageRecord {}