+++ src/redux/message.js
@@ -1,17 +1,11 @@
 // Types
 
 export const types = {
-  CLICKED_DELETE: 'message/CLICKED_DELETE',
   CLICKED_EDIT: 'message/CLICKED_EDIT',
   CLICKED_RESPOND: 'message/CLICKED_RESPOND',
 };
 
 // Action creators
-
-const clickedDelete = messageId => ({
-  type: types.CLICKED_DELETE,
-  payload: { messageId }
-});
 
 const clickedEdit = message => ({
   type: types.CLICKED_EDIT,
@@ -24,7 +18,6 @@ const clickedRespond = message => ({
 });
 
 export const actions = {
-  clickedDelete,
   clickedEdit,
   clickedRespond,
 };